requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.8
django-allauth==0.34.0
django-appconf==1.0.2
django-bootstrap-form==3.3
django-filter==1.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
elasticsearch-dsl==6.2.1
elasticsearch==6.3.0
gunicorn==19.7.1
idna==2.6
Markdown==2.6.9
oauthlib==2.0.6
olefile==0.45.1
Pillow==4.2.1
pinax-messages==2.0.2
pinax-theme-bootstrap==8.0.1
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.5.3
python-logstash==0.4.6
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
whitenoise==3.3.1
